Renamed searchResponse.ts to search.ts
