+++ src/app/shared/services/history-log/history-log.service.ts
@@ -4,7 +4,7 @@ import { Store } from '@ngxs/store';
 import { Observable } from 'rxjs';
 import { ApplicationHistory, FilterData, ProviderAdminHistory, ProviderHistory } from '../../models/history-log.model';
 import { PaginationElement } from '../../models/paginationElement.model';
-import { SearchResponse } from '../../models/searchResponse.model';
+import { SearchResponse } from '../../models/search.model';
 import { PaginatorState } from '../../store/paginator.state';
 
 @Injectable({
