+++ src/app/shell/result/workshop-map-view-list/workshop-map-view-list.component.ts
@@ -13,7 +13,7 @@ import {
 import { Store } from '@ngxs/store';
 import { Observable, Subject } from 'rxjs';
 import { filter, takeUntil } from 'rxjs/operators';
-import { SearchResponse } from '../../../shared/models/searchResponse.model';
+import { SearchResponse } from '../../../shared/models/search.model';
 import { Role } from '../../../shared/enum/role';
 import { Address } from '../../../shared/models/address.model';
 import { PaginationElement } from '../../../shared/models/paginationElement.model';
